GUN
Yay now the bullets move indepently of the player by using a node in the main scene!
Also signals are used to transfer in information about the bullets supposed position direction and rotation to the node, transfered in the signal var("bulletfired",bullet_instance,gunend.global_position,direction)!
jmblv is goated!