Signals and enemies
I think how this works is that you use "enum" to set up 2 diferent states for the enemy to be in, I've choosen LOOK and ATTACK which will change the AI's behaviour. The enemies senses the player through a big detection area that extends a huge distance around an enemy, which aims to detect for the player.