Groups and enemy funkiness and ducks?
The player is added to a group, which allows the bullet to know what its detecting by checking the group the area it hits is in.
The tutorial has an approach using a gun for the enemys, but I'm using a ufo for it to work so I'll have to take a dif approach. I'm wondering if I can copy the code used for summoning the bullets for the ufo, for the player whichuses signals to send data about the location need to spawn the bullet.
Also I just wanted to include duck coding, I felt really sad about not talking about it. Theres this one concept the goat jmbiv talks about, where instead of checking for an bodies ttype you check for its attributes
Basically if it looks like a duck and it acts like a duck and sounds like a duck its probably a duck(but for coding)