GROUPS
So this is where I start on the 2nd week, and to start I'm trying to solve one of the issues i noticed last week: Bullets shot by a ufo would kill a freidnly alien. This is because the bullets for both the player and the ufos were considered the same which caused ufo bullets to act as if shot by the player, so they would kill their friends. I'm trying to remedy this through using a global variable which adds the bullets to a group, changing depending on who shot it. This is really funny because in the chat their fighting over the same variable..
Another problem I had, a tinier one is I did not like the naming system for the scripts, it was really dang confusing trying to figure out what script I wanted to use so I alphabetize them in the one orginization menu, and renamed ONE script from character_body_2d.gd -> player.gd
This was a bit annoying thou, it caused errorrs until I deleted the cache from the folders